# doc/api.yaml
openapi: 3.0.0
info:
  title: WasaProject
  description: The WASA Course Project, building an image sharing platform
  version: 1.0.0
...
security:
  - BearerAuth: []
paths:
  /session:
    post:
      tags: [login]
      summary: Logs in the user
      description: |
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: User details
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string 
                  example: Maria 
                  pattern: 'ˆ.*?$' 
                  minLength: 3 
                  maxLength: 16
        required: true
      responses:
        '201':
          description: User log-in action successful
          content:
            application/json:
              schema:
                type: object 
                properties:
                  identifier:
                  # change here if you decide to use an integer # or any other type of identifier
                    type: string
                    example: "abcdef012345"  
        '400': { $ref: "#/components/responses/BadRequest" }
        '500': { $ref: "#/components/responses/ServerError" }
...
  /users/username/{username}:
    parameters:
        - name: username
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/username'
    patch:
      tags: [user]
      summary: Set My User Name
      description: Set the user's username.
      operationId: setMyUserName
      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'

      responses:
        '200':
          description: action successful
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Success'

        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/ServerError" }
  
    get:
      tags: [user]
      summary: Get User Profile
      description: Get the profile of a user.
      operationId: getUserProfile
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/ServerError" }
  /users/id/{userID}:
    parameters:
        - name: userID
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/user'
    get:
      tags: [user]
      summary: Get User Profile
      description: Get the profile of a user.
      operationId: getUserProfile
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/ServerError" }
  /users/{username}/followers:
    parameters:
        - name: username
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/username'
    post:
      tags: [user]
      summary: Follow User
      description: Follow a user.
      operationId: followUser
      responses:
        '200':
          description: action successful
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Success'

        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/ServerError" }

    delete:
      tags: [user]
      summary: Unfollow User
      description: Unfollow a user.
      operationId: unfollowUser
      responses:
        '200':
          description: action successful
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Success'

        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/ServerError" }

    get:
      tags: [user]
      summary: Get Followers
      description: Get the followers of a user.
      operationId: getFollowers
      responses:
        '200':
          description: Followers retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
                  minItems: 0
                  maxItems: 100

        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/ServerError" }
...
components:
  responses:
    BadRequest:
      description: Error Code 400
    Unauthorized:
      description: Error Code 401
    ServerError: 
      description: Error Code 500
  schemas:
    Success: 
      type: string
      minLength: 1
      maxLength: 100
      pattern: '^[a-zA-Z0-9]$'
      description: A message indicating the success of an operation
...
    username:
      type: string
      minLength: 10
      maxLength: 20
      pattern: "^[a-zA-Z0-9_]+$"
      description: A unique user identifier
...
    user: 
      type: object
      properties:
        userId:
          type: string
          minLength: 10
          maxLength: 20
          pattern: "^[a-zA-Z0-9_]+$"
          description: A unique user identifier
        username:
          type: string
          minLength: 3
          maxLength: 50
          pattern: "^[a-zA-Z0-9_]+$"
          description: The user's username
        followers:
          type: array
          items:
            type: string
            description: Unique identifier of a follower.
            minLength: 10
            maxLength: 20
            pattern: "^[a-zA-Z0-9_]+$"
          description: An array of unique identifiers for the user's followers.
        following:
          type: array
          items:
            type: string
            description: Unique identifier of a followed user.
            minLength: 10
            maxLength: 20
            pattern: "^[a-zA-Z0-9_]+$"
          description: An array of unique identifiers for the users being followed by the user.
        photos:
          type: array
          items:
            type: string
            description: Unique identifier of a photo uploaded by the user.
            minLength: 10
            maxLength: 20
            pattern: "^[a-zA-Z0-9_]+$"
          description: An array of unique identifiers for the photos uploaded by the user.
      required:
        - userId
        - username
      description: Represents a user, including information about their followers, who they're following, and their photos.
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT 